api/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0
requests==2.31.0
psutil==5.9.6
langchain==0.0.350
langchain-community==0.0.10
langchain-groq==0.0.3
langgraph==0.0.20
pytrends==4.9.2
groq==0.4.1